+++ src/gameObject/GoalTop.cpp
@@ -1,6 +1,6 @@
 #include "gameObject/GoalTop.h"
 
-GoalTop::GoalTop(int x, int y, bool needScale)
+GoalTop::GoalTop(float x, float y, bool needScale)
 {
 	m_sprite.setTexture(Resources::getInstance().getGoalTexture(2));
 	m_sprite.setPosition(x, y);
@@ -8,17 +8,17 @@ GoalTop::GoalTop(int x, int y, bool needScale)
 
 	if (needScale) {
 
-		m_sprite.scale(-1, 1);
+		m_sprite.scale(-1.f, 1.f);
 	}
 }
 
 
 bool GoalTop::m_registeritLeftGoalTop = StaticFactory::registeritStatic("LeftGoalTop",
-	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(40, 580, false); });
+	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(40.f, 580.f, false); });
 
 
 bool GoalTop::m_registeritRightGoalTop = StaticFactory::registeritStatic("RightGoalTop",
-	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(1750, 580, true); });
+	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(1750.f, 580.f, true); });
 
 void GoalTop::draw(sf::RenderWindow& window)const {
